requirements.txt
torchvision==0.4.0
numpy==1.17.2
torch==1.2.0
opencv_python==4.1.0.25
tensorboardX==1.8
Pillow==5.1.0
pthflops==0.2.2
